Avoid re-uploading images when possible
Every time a function taking an image as an input is called, Gradio
uploads the image to the server.  For instances running in the cloud
with a slow uplink, this is a major performance bottleneck.

Gradio doesn't allow caching images, but this can be worked around by
calculating the image checksum, using the checksum as the actual
function input, and only uploading the image the first time it is
selected.