Fix missing parameters in browser-displayed images
Processed images are passed to Gradio as PIL images which don't contain
generation parameters.  This means that when a user saves an image by
right-clicking in the browser, they end up with a file that is almost
identical to the processed file but missing the parameter chunk.

Apart from being a bug by itself, this means that when a user selects a
generated image as input for img2img/inpaint, it has to be re-uploaded
because the UI doesn't recognize the file.

This commit solves both problems by saving processed images to a file,
creating a temporary file if the image wouldn't have been saved
otherwise, and passing the filename to Gradio.