Fix missing parameters in browser-displayed images
Processed images are passed to Gradio as PIL images which don't contain
generation parameters.  This means that when a user saves an image by
right-clicking in the browser, they end up with a file that is almost
identical to the processed file but missing the parameter chunk.

This commit solves this by saving processed images to a file, creating
a temporary file if the image wouldn't have been saved otherwise, and
passing the filename to Gradio.