Implement multi-cond guidance for Composable Diffusion.
As a bonus, add prompt weighting too.

Based on https://arxiv.org/pdf/2206.01714.pdf /
https://energy-based-model.github.io/Compositional-Visual-Generation-with-Composable-Diffusion-Models/

In vanilla Stable Diffusion, prompt generations is guided based on two
prompts: towards the positive prompt, and away from the negative prompt.
This change allows you to use an arbitrary number of prompts for
guidance, for some interesting composition options. See the website
above for more concrete examples.

Multi-cond guidance slows generation because it requires evaluating
guidance for additional prompts for each step.

New syntax keywords: AND NOT PLUS -- since CLIP is case insensitive,
just have them lowercase to use them in a prompt.

SYNTAX GUIDE:

Watch the console for debugging output of how each prompt is evaluated.

New case-sensitive keywords: AND NOT PLUS. Weights are :NUMBER.

"red AND white" guides with a "red" prompt and a "white" prompt.

"red:2 AND white" guides with a "red" prompt 2x stronger than a "white" prompt

"a photo of a (cat AND dog)" is equivalent to "a photo of a cat AND a
photo of a dog" and generate an animal hybrid using the two prompts.

"a person NOT human" guides towards "a person" with "human" as a prompt
with -0.5x weight.

"cat PLUS dog" guides with a prompt made by adding the CLIP-embeddings
from "cat" to "dog" and dividing by 2.

You can combine PLUS and AND. "apple PLUS pear AND banana PLUS eggplant"
makes an image containing apple/pear hybrids and banana/eggplant
hybrids.

Multiple paren groups are supported and combine groups sensibly.
"photo of (dog AND cat), cute, 4k, playing with (ball AND yarn)"
=> "photo of dog, cute, 4k, playing with ball AND photo of cat, cute,
4k, playing with yarn".