Make it so we can load checkpoints from "pasted" params
I know, it's a lot. But I had to shuffle things around and move declarations to other places and rename stuff so it was all unique, as I'm sure we don't want to load params to both txt2img and img2img.

The alternative is to just make another row at the top with the checkpoint selection, which can then be referenced without all the shuffling around.