Added "Comment" function to negative prompt
This adds following functions (to the negative prompt only):
1. You can use "#" before one or several words and it will delete those until the next comma (f.e.: "#blue, round, cubic" --> "round, cubic")
2. You can use "##" to everything behind those characters (f.e.: "blue, ##round, cubic" --> "blue")
3. You can use "\#" or "\##" to simply undo the comment function (f.e.: "\#blue, \##round, cubic" --> "blue, round, cubic" )
This might not be useful to everyone but it can save a lot of time when messing with the negative prompt because you don't have to crop your prompt to somewhere only to paste it back later. It's also useful to add a "standard" negative prompt in the config which can be uncommented at any time when needed.