Refactored opt.apply functions for samplers funcs
apply_sampler should not build or check for errors, it should
only apply the sampler to the process handed to it.

process_axis, when axis was "Samplers", did not finish processing
the valslist. For this, I implemented build_samplers_valslist.

build_samplers_dict was being called twice, once in apply, and once
in confirm_samplers, I pulled this out from confirm and put it into
an if opt.label == samplers block in process_axis

build_samplers_dict was modified to return a dictionary with a different
structure.
I believe this new structure makes it more clear for readers.
The previous structure was a dictionary where the first key was the
lowercase version of sampler.name, and adjacent was a key that was an
alias, both set with the same int value, for all samplers.
Because dictionaries are now ordered, dict.get(val) would return the first
match.
Also, this structure did not include the proper name needed by
Process.sampler_name. The structure is now
    {Sampler.name: [Sampler.name.lower(), alias1, ...], }
Now you can check if user value is in set of flattened dict.values(),
    "set(chain(*dict.values()))"
and grab Sampler.name from the key directly.

The building of valslist, utilizes this structure, function is
    build_samplers_valslist
It iterates through (previously confirmed) user supplied values,
steps through each dict.values until it matches, then appends
the proper name from key into list