rudimentary aspect ratios
Adds a ui setting where a user can specify their
preferred aspect ratios.
Uses this to generate the buttons for txt2img &
img2img, with element ids.
To keep the user experience intuitive, the ratio uses colon format
":" to separate width and height.
This value is used as the button value, parsed as ints for the ratio
value, and used in the element id.