Improved instructions
Improved instructions with proper indications, the sentence: "One click install and run script (but you still must install python and git)" it means is not clear, if it's just a one click install? should be included the Git and Python setup steps on the bash script also? To allow more users to adopt this technology and respect the author, I just proper indicate where to download both softare and do not touch the bash script install.

relate with Improved instructions #8357 issue