remove unnecessary batch_size designation
I was trying to suppress showing image in web when running a xyz batch, but I found I could not adjust the batch size. 
The code block I touched will set batch_size to 1 if Settings->User interface->"Show grid in results for web" is unchecked (opts.return_grid=False). This preset does not make sense because the originating setting is for UI image visualization, not for the image production. 

I am not familiar with other parts of the project so please point out any potential side effect.