Update img2img.py
The `process_batch` function takes a `to_scale` parameter, but it does not check whether this parameter is actually a boolean. If the parameter is not a boolean, then the code will try to convert it to a float, which could cause an error. This code will check whether the `to_scale` parameter is a boolean. If it is not, then the code will raise a `TypeError` exception.