Escape HTML in user metadata description before appending to card
Currently the description is appended to the document as-is. This
allows for the user to inadvertently add HTML to the model
description shown on-card and have it be appended onto the DOM if
`show descriptions on card` is enabled. On a single-user system
this presents a minimal security hazard, but on a shared system,
any user could insert malicious code into the card metadata via
the metadata editor.