Update launch_utils.py
When the commithash is given, we only clone the needed version. When the commithash is not specified, we only clone the latest version. This is to save time and network traffic as well as some disk space, since as time passes by, those repos can be big, and for those who have difficulty using github (like me, and many other developers in China), this can be convenient  to us.