Update webui.sh
This enhancement introduces a crucial check for the existence of the 'stable-diffusion-webui' folder within the project directory. The motivation behind this addition is to mitigate potential conflicts when the project is situated within another Git repository, a situation often encountered when users operate live servers with their own Git configurations.

Previously, the absence of this check could lead to unintended interactions with the enclosing Git repository, resulting in unforeseen issues. Now, by verifying the presence of the 'stable-diffusion-webui' folder, the script ensures a more resilient and isolated deployment. If the folder exists, it is leveraged as the install directory, guaranteeing a seamless integration while minimizing the risk of interference with surrounding Git repositories.

This improvement aims to elevate compatibility and prevent unintended consequences, especially in scenarios where the project coexists within other Git-controlled environments.