Upgrade PIP during virtual environment creation
The default version of PIP with Python 3.10 does not record installation metadata properly for requirements installed using an url, resulting in invalid lines when PIP FREEZE is called.  For example, the packages clip and cstr are problematic.  This makes it difficult to recreate a stable diffusion virtual environment.  Upgrading PIP during virtual environment creation remedies this problem.  I have found this is corrected in v24 of PIP.